test_util.py
"""
This test module evaluates the functionality of the 'util' module.

The 'util' module provides functions for file handling.

Functions Tested:
- detect_file_type(fn): Determines the file type of a given file using the 'file' command.
- read_db(db_file, args={'encoding':'utf-8'}):
	Reads a file into a Pandas DataFrame based on its format.

Test Cases:
1. test_detect_file_type_existing: Tests the 'detect_file_type' function
	with an existing file. It checks whether the function correctly
    identifies the file type.

4. test_detect_file_type_existing_csv: Tests the 'detect_file_type'
    function with an existing CSV file. Similar to 'test_detect_file_type_existing',
    this test case checks whether the function correctly identifies the file type of an existing CSV file.
5. test_detect_file_type_nonexisting: Tests the 'detect_file_type' function 
    with a non-existing file. Similar to 'test_read_db_nonexisting',
    this test case verifies that the function 
    returns an empty string when given the path to a non-existing file.

Test Fixture:
- setUp: Creates a temporary CSV file for testing before each test case is executed.
- tearDown: Removes the temporary CSV file after each test case is executed.

Usage:
To run the test suite, execute this module.
The 'unittest.main()' call at the end of the module triggers the test execution.

Note:
- These tests are designed to ensure the accuracy and reliability of
	the functions provided by the 'util' module.
- It's important to have a clean test environment to ensure 
	the integrity of the test results.

For further details about the 'util' module and its functions,
refer to the module's docstring and implementation.
"""

import unittest
import os
import pandas as pd
from util import detect_file_type


class TestParseRow(unittest.TestCase):
    """
    Test suite for the 'read_db' module.
    """

    def setUp(self):
        # Create a temporary CSV file for testing
        self.test_csv_path = 'test.csv'
        with open(self.test_csv_path, 'w', encoding='utf-8') as f:
            f.write('Test data\n')

    def tearDown(self):
        # Remove the temporary CSV file after testing
        if os.path.exists(self.test_csv_path):
            os.remove(self.test_csv_path)

    def test_detect_file_type_existing(self):
        """
        Test the detect_file_type function with an existing file.
        """
        detected_type = detect_file_type(self.test_csv_path)
        self.assertEqual(detected_type, 'text/plain')

    def test_detect_file_type_existing_csv(self):
        """
        Test the detect_file_type function with an existing CSV file.
        """
        # Create a temporary CSV file
        test_csv_path = 'test.csv'
        with open(test_csv_path, 'w', encoding='utf-8') as f:
            f.write('Test data\n')

        # Call the detect_file_type function
        detected_type = detect_file_type(test_csv_path)

        # Assert that the detected type matches the expected type
        self.assertEqual(detected_type, 'text/plain')

        # Remove the temporary CSV file
        os.remove(test_csv_path)

    def test_detect_file_type_nonexisting(self):
        """
        Test the detect_file_type function with a non-existing file.
        """
        # Call the detect_file_type function with a non-existing file
        detected_type = detect_file_type('non_existing.csv')

        # Assert that an empty string is returned
        self.assertEqual(detected_type, '')


if __name__ == '__main__':
    unittest.main()
